1.25.1 - Taming Twitter
Fixes #13
Up until now only the first mutation per node was scrutinised.

I also increased the depth up the tree that we check when looking if we're inside a contentEditable element.

And manually localised times now must have either a separator or a meridiem for it to be localised. This fixes random numbers being wrongly localised :)