1.40.3
Re-inject the page content script when the extension updates. (This is a change for Manifest V3/Chrome)
This should resolve the issue of the popup not functioning properly on pages after an update, however it does require the "scripting" permission.
Avoid needlessly sending the enabled state to tabs other than the active one when changing the state.